+++ src/services/domain/organizations.service.ts
@@ -1,7 +1,9 @@
-import { createAnonymousClient } from "@/utils/supabase/anonymousClient"
-import Bucket from "@/utils/supabase/bucket"
-import { createClient } from "@/utils/supabase/server"
-import { NextResponse } from "next/server"
+import FileUtils from '@/utils/files/fileUtils'
+import { createAnonymousClient } from '@/utils/supabase/anonymousClient'
+import Bucket from '@/utils/supabase/bucket'
+import { createClient } from '@/utils/supabase/server'
+import { revalidatePath } from 'next/cache'
+import { NextResponse } from 'next/server'
 
 const ORGANIZATIONS = 'organizations'
 
@@ -22,4 +24,36 @@ export async function getOrganizations() {
   )
 
   return NextResponse.json(organizationsWithLogos)
+}
+
+export async function postOrganization(newOrganization: any) {
+  const supabaseClient = createAnonymousClient()
+  const supabaseClientAuth = createClient()
+  const user = await supabaseClientAuth.auth.getUser()
+  const idUser = user.data.user?.id
+
+  const logoSrc = newOrganization.logoSrc
+  delete newOrganization.logoSrc
+
+  const { data, error } = await supabaseClient
+    .from(ORGANIZATIONS)
+    .insert([{ ...newOrganization, created_by: idUser }])
+    .select()
+  if (error) return NextResponse.json(error, { status: 500 })
+
+  const organizationId = data[0].id
+
+  try {
+    const extension = FileUtils.getFileExtensionFromBase64(logoSrc)
+    const logoPath = `organizations/logos/${organizationId}.${extension}`
+    await Bucket.uploadFile(logoPath, logoSrc)
+    await supabaseClient.from(ORGANIZATIONS).update({ logoPath }).eq('id', organizationId)
+  } catch (error) {
+    await supabaseClient.from(ORGANIZATIONS).delete().eq('id', organizationId)
+    return NextResponse.json(error, { status: 500 })
+  }
+
+  revalidatePath('/api/organizations')
+
+  return NextResponse.json(data[0])
 }
